Refactor EngagementViewController environment initialization
